refactor: collection
