+++ src/features/quiz/config/index.tsx
@@ -1,0 +1,47 @@
+export const quizzes: Quiz.List = [
+  {
+    id: '1',
+    type: 'multiple',
+    question: '식물기반 단백질 시장에서 대기업의 참여가 늘어나는 이유는 무엇인가요?',
+    options: [
+      '기존의 배양육이 기존방식에서 생산되는 육류보다 토양이 비축된다',
+      '배양육이 기존방식에서 생산되는 육류보다 토양이 비축된다',
+      '기존의 배양육이 기존방식에서 생산되는 육류보다 토양이',
+      '기존의 배양육이 기존방식에서 육류보다 토양이 비축된다',
+    ],
+    answer: '기존의 배양육이 기존방식에서 생산되는 육류보다 토양이 비축된다',
+    explanation: '기존의 배양육이 기존방식에서 생산되는 육류보다 토양이 비축된다',
+  },
+  {
+    id: '2',
+    type: 'ox',
+    question: '식물기반 단백질 시장에서 대기업의 참여가 늘어나는 이유는 무엇인가요?',
+    answer: 'O',
+    explanation: '기존의 배양육이 기존방식에서 생산되는 육류보다 토양이 비축된다',
+  },
+  {
+    id: '3',
+    type: 'multiple',
+    question: '식물기반 단백질 시장에서 대기업의 참여가 늘어나는 이유는 무엇인가요?',
+    options: [
+      '기존의 배양육이 기존방식에서 생산되는 육류보다 토양이 비축된다',
+      '배양육이 기존방식에서 생산되는 육류보다 토양이 비축된다',
+      '기존의 배양육이 기존방식에서 생산되는 육류보다 토양이',
+      '기존의 배양육이 기존방식에서 육류보다 토양이 비축된다',
+    ],
+    answer: '기존의 배양육이 기존방식에서 생산되는 육류보다 토양이 비축된다',
+    explanation: '기존의 배양육이 기존방식에서 생산되는 육류보다 토양이 비축된다',
+  },
+  {
+    id: '4',
+    type: 'ox',
+    question: '식물기반 단백질 시장에서 대기업의 참여가 늘어나는 이유는 무엇인가요?',
+    answer: 'O',
+    explanation: '기존의 배양육이 기존방식에서 생산되는 육류보다 토양이 비축된다',
+  },
+]
+
+export const tabs = [
+  { key: 'exploration', label: '탐색' },
+  { key: 'my-collection', label: '내 컬렉션' },
+] as const
